stm32: Implement manual double buffering tx for usbotg
It is common for USB host controllers to send back-to-back IN tokens
which only gives the MCU ~3us to queue the next USB packet in the
hardware.  That can be difficult to do if the MCU has to wake up the
task code.  Add support for queuing the next packet destined for the
host so that the USB irq handler can directly load it into the
hardware.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>